Extract bingo card into separate component #132  #92  #35
Extracted the bingo card into a separate component and moved all related logic from page.svelte into it. The page.svelte file is now clean and more maintainable. #35 #92  #132
